Split lookup decompression to a separate file
